# app/rest/docs/swagger.yaml
basePath: /api/v1
definitions:
  rest.AddTagRequest:
    properties:
      tag_id:
        type: string
    type: object
  rest.CreateGuestRequest:
    properties:
      doc:
        type: string
      name:
        type: string
    type: object
  rest.CreateScoreRequest:
    properties:
      date:
        type: string
      description:
        type: string
      guest_id:
        type: string
    type: object
  rest.CreateTagRequest:
    properties:
      name:
        type: string
    type: object
  rest.Guest:
    properties:
      created_at:
        type: string
      doc:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  rest.HTTPResponse:
    properties:
      msg:
        example: any message
        type: string
    type: object
  rest.IDResponse:
    properties:
      id:
        type: string
    type: object
  rest.Score:
    properties:
      created_at:
        type: string
      doc:
        type: string
      id:
        type: string
      name:
        type: string
      tags:
        items:
          $ref: '#/definitions/rest.Tag'
        type: array
      updated_at:
        type: string
      used_in:
        type: string
      was_used:
        type: boolean
    type: object
  rest.Tag:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact:
    email: oi@coding4u.com.br
    name: Coding4u
  description: Swagger API for Loyalty Card Service.
  termsOfService: http://swagger.io/terms/
  title: Loyalty Card Swagger API
  version: "1.0"
paths:
  /guests:
    post:
      consumes:
      - application/json
      description: Router for create a new guest
      operationId: createGuest
      parameters:
      - description: JSON body for create a new guest
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/rest.CreateGuestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.IDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.HTTPResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest.HTTPResponse'
      summary: create a new guest
      tags:
      - Guest
  /guests/{guest_id}:
    get:
      consumes:
      - application/json
      description: Router for find a guest
      operationId: findGuest
      parameters:
      - description: Guest ID
        in: path
        name: guest_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.Guest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.HTTPResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest.HTTPResponse'
      summary: find a gust
      tags:
      - Guest
  /scores:
    post:
      consumes:
      - application/json
      description: Router for create a new score
      operationId: createScore
      parameters:
      - description: JSON body for create a new score
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/rest.CreateScoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.IDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.HTTPResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest.HTTPResponse'
      summary: create a new score
      tags:
      - Score
  /scores/{score_id}:
    get:
      consumes:
      - application/json
      description: Router for find a score
      operationId: findScore
      parameters:
      - description: Score ID
        in: path
        name: score_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.Score'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.HTTPResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest.HTTPResponse'
      summary: find a score
      tags:
      - Score
  /scores/{score_id}/tags:
    post:
      consumes:
      - application/json
      description: Router for add a tag to a score
      operationId: addTag
      parameters:
      - description: Score ID
        in: path
        name: score_id
        required: true
        type: string
      - description: JSON body for add a tag to a score
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/rest.AddTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.HTTPResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.HTTPResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest.HTTPResponse'
      summary: add a tag to a score
      tags:
      - Score
  /scores/{score_id}/use:
    post:
      consumes:
      - application/json
      description: Router for use a score
      operationId: useScore
      parameters:
      - description: Score ID
        in: path
        name: score_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.HTTPResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.HTTPResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest.HTTPResponse'
      summary: use a score
      tags:
      - Score
  /tags:
    get:
      consumes:
      - application/json
      description: Router for search tags
      operationId: searchTags
      parameters:
      - in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/rest.Tag'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.HTTPResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest.HTTPResponse'
      summary: search tags
      tags:
      - Tag
    post:
      consumes:
      - application/json
      description: Router for create a new tag
      operationId: createTag
      parameters:
      - description: JSON body for create a new tag
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/rest.CreateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.IDResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.HTTPResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest.HTTPResponse'
      summary: create a new tag
      tags:
      - Tag
  /tags/{tag_id}:
    get:
      consumes:
      - application/json
      description: Router for find a tag
      operationId: findTag
      parameters:
      - description: Tag ID
        in: path
        name: tag_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.Tag'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.HTTPResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/rest.HTTPResponse'
      summary: find a tag
      tags:
      - Tag
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"